PycharmProjects/logbit-assessment-master/logbit-assessment-master/requirements-dev.txt
asgiref==3.3.4
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.4
django-appconf==1.0.5
django-avatar==5.0.0
django-colorfield==0.2.1
django-crum==0.7.9
django-filter==2.4.0
django-import-export==2.7.1
django-nested-admin==3.2.3
django-object-actions==3.0.1
django-quill-editor==0.1.38
django-relativedelta==1.1.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.9
pathspec==0.9.0
Pillow==8.1.2
platformdirs==2.4.0
python-decouple==3.4
python-monkey-business==1.0.0
pytz==2021.1
PyYAML==3.10
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tablib==3.1.0
tomli==1.2.3
typing-extensions==3.10.0.0
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
